requirements.txt
# Core FastAPI
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Configuration
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Providers
openai==1.57.2
google-generativeai==0.8.3

# SSE Streaming
sse-starlette==2.2.1

# Async File I/O
aiofiles==24.1.0

# HTTP Client
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
